chore(cli): remove the aide daemon command
`aide daemon` served the same gRPC API as the MCP primary from a strictly
weaker process: bolt without bleve, so memory search fell back to an
unranked substring scan; no instinct store, so those tools were dead for
every attached client; no observe sink or event bus; no watcher, no startup
reconcile, no findings analysers.

Both claim the same socket, so this was not merely a lesser option — a
daemon started first won the socket and every later `aide mcp` dropped into
client mode behind it, silently inheriting all of the above. Someone had
already patched one symptom of that (the SetCodeReconciler hook, added
because reconcileCode() was otherwise a no-op).

Nothing launched it: no unit file, no supervisor, no caller anywhere in the
tree. Its only appearances were three doc lines and the dispatch arm.

The retention machinery it happened to house is shared with the MCP primary
and the direct-mode sweep at session init, so cmd_daemon.go becomes
retention.go with just that.

"daemon" stays as the generic name for whatever holds the socket, which is
how registry, backend.go and aide-web already use it — unambiguous now that
no command answers to it.
